displayAll

+++ src/main/java/com/netcracker/repository/BoxRepository.java
@@ -4,7 +4,7 @@ import com.netcracker.model.Box;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface BoxRepository extends JpaRepository<Box, Long> {
+public interface BoxRepository extends JpaRepository<Box, Integer> {
 
 
 }
